requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.1.3
openpyxl==3.1.2
Pillow==10.1.0

# OCR
pytesseract==0.3.10
pdf2image==1.16.3

# Vector DB and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# Search
whoosh==2.7.4  # For keyword search

# LLM
openai==1.3.7
langchain==0.1.0

# Utilities
python-magic-bin==0.4.14
langdetect==1.0.9
aiofiles==23.2.1


